+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Release 5.11.0 (unreleased):
    - In `SimpleAuthorizationService` add parameter `configurationIds` to method `credentialOfferWithAuthorizationCode`
    - Support different supported credential formats having the same scope value (as this is covered by the spec)
    - Be sure to use correct JWE encryption algorithm on issuer and wallet side and advertise keys correctly
+   - In `WalletEncryptionService` add constructor parameter `fallbackJweEncryptionAlgorithm` and deprecated `supportedJweEncryptionAlgorithm`
  - OpenID for Verifiable Presentations:
    - In `OpenId4VpVerifier` add constructor parameter `supportedJweEncryptionAlgorithms` to advertise in metadata, deprecating `supportedJweEncryptionAlgorithm` and `supportedJweAlgorithm`
  - OAuth 2.0: